cyto-plot/pyproject.toml
[project]
name = "cyto-plot"
version = "0.1.0"
description = "Plot cytometry data from FCS files using UMAP and Bokeh"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh>=3.7.3",
    "cudf-cu12>=25.8.0",
    "cuml-cu12>=25.8.0",
    "joblib>=1.5.1",
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "readfcs>=2.0.1",
    "ruff>=0.12.8",
]
